конфликт

+++ src/SalesManager.java
@@ -13,9 +13,9 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale : sales) {
+    public long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale : sales) {
             if (sale < min) {
                 min = sale;
             }
@@ -23,9 +23,9 @@ public class SalesManager {
         return min;
     }
 
-    public int stat(){
-        int stat = 0;
-        for (int sale : sales){
+    public long stat(){
+        long stat = 0;
+        for (long sale : sales){
             stat += sale;
             }
         return (stat - min()-max())/(sales.length - 2);
